requirements.txt
pandas 
numpy
python-dotenv
pymongo
certifi
pymongo[srv]==3.11
scikit-learn
pyyaml>=6.0

-e .